products/pyproject.toml
[project]
name = "products"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Cully Wakelin", email = "cully.wakelin@resonatehq.io" }
]
dependencies = [
    "resonate-sdk==0.4.8",
    "colorlog>=6.9.0",
]
readme = "README.md"
requires-python = ">= 3.9"

[project.scripts]
"products" = "products:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "mypy>=1.14.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/products"]
